Move old implementation of arbitrary state and utils into another branch
